Integrated SSE decoder into pdsch receiver

+++ srslte/include/srslte/fec/rm_turbo.h
@@ -77,8 +77,8 @@ SRSLTE_API int srslte_rm_turbo_rx(float *w_buff,
                                   uint32_t rv_idx, 
                                   uint32_t nof_filler_bits);
 
-SRSLTE_API int srslte_rm_turbo_rx_lut(float *input, 
-                                      float *output, 
+SRSLTE_API int srslte_rm_turbo_rx_lut(int16_t *input, 
+                                      int16_t *output, 
                                       uint32_t in_len, 
                                       uint32_t cb_idx, 
                                       uint32_t rv_idx);
